+++ crates/everruns-worker/src/lib.rs
@@ -6,19 +6,9 @@ pub mod runner_inprocess;
 pub mod tools;
 pub mod workflows;
 
-// Temporal-specific modules (feature-gated)
+// Temporal-specific module (feature-gated)
 #[cfg(feature = "temporal")]
-pub mod activities_temporal;
-#[cfg(feature = "temporal")]
-pub mod runner_temporal;
-#[cfg(feature = "temporal")]
-pub mod temporal_client;
-#[cfg(feature = "temporal")]
-pub mod temporal_types;
-#[cfg(feature = "temporal")]
-pub mod temporal_worker;
-#[cfg(feature = "temporal")]
-pub mod workflows_temporal;
+pub mod temporal;
 
 // Re-export main types
 pub use executor::WorkflowExecutor;
