+++ crates/everruns-worker/src/runner.rs
@@ -108,8 +108,19 @@ pub async fn create_runner(
                 "Creating Temporal agent runner"
             );
 
-            let runner = crate::runner_temporal::TemporalRunner::new(config.clone(), db).await?;
-            Ok(Arc::new(runner))
+            #[cfg(feature = "temporal")]
+            {
+                let runner = crate::temporal::TemporalRunner::new(config.clone(), db).await?;
+                Ok(Arc::new(runner))
+            }
+
+            #[cfg(not(feature = "temporal"))]
+            {
+                anyhow::bail!(
+                    "Temporal runner requested but 'temporal' feature is not enabled. \
+                    Compile with: cargo build --features temporal"
+                )
+            }
         }
     }
 }
